+++ R/cpp11.R
@@ -596,6 +596,10 @@ distribute_tool_to_set_cpp <- function(agents_ids) {
   .Call(`_epiworldR_distribute_tool_to_set_cpp`, agents_ids)
 }
 
+distribute_tool_to_entities_cpp <- function(prevalence, as_proportion) {
+  .Call(`_epiworldR_distribute_tool_to_entities_cpp`, prevalence, as_proportion)
+}
+
 virus_cpp <- function(name, prevalence, as_proportion, prob_infecting, prob_recovery, prob_death, post_immunity, incubation) {
   .Call(`_epiworldR_virus_cpp`, name, prevalence, as_proportion, prob_infecting, prob_recovery, prob_death, post_immunity, incubation)
 }
@@ -686,4 +690,8 @@ distribute_virus_randomly_cpp <- function(prevalence, as_proportion, agents_ids)
 
 distribute_virus_to_set_cpp <- function(agents_ids) {
   .Call(`_epiworldR_distribute_virus_to_set_cpp`, agents_ids)
+}
+
+distribute_virus_to_entities_cpp <- function(prevalence, as_proportion) {
+  .Call(`_epiworldR_distribute_virus_to_entities_cpp`, prevalence, as_proportion)
 }